remove server module
