+++ helix-rest/pom.xml
@@ -46,8 +46,8 @@ under the License.
   <dependencies>
     <dependency>
       <groupId>org.apache.helix</groupId>
-      <artifactId>metadata-store-directory-server</artifactId>
-      <version>${project.version}</version>
+      <artifactId>metadata-store-directory-common</artifactId>
+      <version>0.9.2-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>org.yaml</groupId>
