--- metadata-store-directory-server/src/test/resources/log4j.properties
@@ -1,41 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-# Set root logger level to DEBUG and its only appender to R.
-log4j.rootLogger=ERROR, C
-
-# A1 is set to be a ConsoleAppender.
-log4j.appender.C=org.apache.log4j.ConsoleAppender
-log4j.appender.C.layout=org.apache.log4j.PatternLayout
-log4j.appender.C.layout.ConversionPattern=%-4r [%t] %-5p %c %x - %m%n
-
-log4j.appender.R=org.apache.log4j.RollingFileAppender
-log4j.appender.R.layout=org.apache.log4j.PatternLayout
-log4j.appender.R.layout.ConversionPattern=%5p [%C:%M] (%F:%L) - %m%n
-log4j.appender.R.File=target/ClusterManagerLogs/log.txt
-
-log4j.appender.STATUSDUMP=org.apache.log4j.RollingFileAppender
-log4j.appender.STATUSDUMP.layout=org.apache.log4j.SimpleLayout
-log4j.appender.STATUSDUMP.File=target/ClusterManagerLogs/statusUpdates.log
-
-log4j.logger.org.I0Itec=ERROR
-log4j.logger.org.apache=ERROR
-log4j.logger.com.noelios=ERROR
-log4j.logger.org.restlet=ERROR
-
-log4j.logger.org.apache.helix.monitoring.ZKPathDataDumpTask=ERROR,STATUSDUMP
